use parseInt instead of parseFloat

+++ js/map.js
@@ -1,4 +1,4 @@
-var size = parseFloat($("#map").height());
+var size = parseInt($("#map").height());
 console.log(size);
 
 var svg = d3.select("div#map").append("svg")
